fix_: create request logger ad-hoc in tests
Fixes `TestCall` failing when run concurrently.
