+++ mobile/callog/status_request_log.go
@@ -1,4 +1,4 @@
-package statusgo
+package callog
 
 import (
 	"fmt"
@@ -11,7 +11,6 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/status-im/status-go/logutils"
-	"github.com/status-im/status-go/logutils/requestlog"
 )
 
 var sensitiveKeys = []string{
@@ -47,7 +46,7 @@ func getShortFunctionName(fn any) string {
 	return parts[len(parts)-1]
 }
 
-// call executes the given function and logs request details if logging is enabled
+// Call executes the given function and logs request details if logging is enabled
 //
 // Parameters:
 //   - fn: The function to be executed
@@ -59,10 +58,10 @@ func getShortFunctionName(fn any) string {
 // Functionality:
 // 1. Sets up panic recovery to log and re-panic
 // 2. Records start time if request logging is enabled
-// 3. Uses reflection to call the given function
+// 3. Uses reflection to Call the given function
 // 4. If request logging is enabled, logs method name, parameters, response, and execution duration
 // 5. Removes sensitive information before logging
-func call(fn any, params ...any) any {
+func Call(logger *zap.Logger, fn any, params ...any) any {
 	defer func() {
 		if r := recover(); r != nil {
 			logutils.ZapLogger().Error("panic found in call", zap.Any("error", r), zap.Stack("stacktrace"))
@@ -72,7 +71,7 @@ func call(fn any, params ...any) any {
 
 	var startTime time.Time
 
-	requestLoggingEnabled := requestlog.IsRequestLoggingEnabled()
+	requestLoggingEnabled := logger != nil
 	if requestLoggingEnabled {
 		startTime = time.Now()
 	}
@@ -102,7 +101,7 @@ func call(fn any, params ...any) any {
 		paramsString := removeSensitiveInfo(fmt.Sprintf("%+v", params))
 		respString := removeSensitiveInfo(fmt.Sprintf("%+v", resp))
 
-		requestlog.GetRequestLogger().Debug("call",
+		logger.Debug("call",
 			zap.String("method", methodName),
 			zap.String("params", paramsString),
 			zap.String("resp", respString),
@@ -113,8 +112,8 @@ func call(fn any, params ...any) any {
 	return resp
 }
 
-func callWithResponse(fn any, params ...any) string {
-	resp := call(fn, params...)
+func CallWithResponse(logger *zap.Logger, fn any, params ...any) string {
+	resp := Call(logger, fn, params...)
 	if resp == nil {
 		return ""
 	}
